Chord and Velocity Keyboards display CANT when entering scale

+++ src/deluge/gui/ui/keyboard/keyboard_screen.cpp
@@ -407,8 +407,10 @@ ActionResult KeyboardScreen::buttonAction(deluge::hid::Button b, bool on, bool i
 
 	// Scale mode button
 	if (b == SCALE_MODE) {
-		if (getCurrentOutputType() == OutputType::KIT) {
-			return ActionResult::DEALT_WITH; // Kits can't do scales!
+		if ((getCurrentOutputType() == OutputType::KIT) || (getCurrentInstrumentClip()->keyboardState.currentLayout == KeyboardLayoutType::KeyboardLayoutTypeChords)) {
+			// Kits and Chords can't do scales!
+			display->displayPopup(deluge::l10n::get(deluge::l10n::String::STRING_FOR_KEYBOARD_VIEW_CANT_ENTER_SCALE));
+			return ActionResult::DEALT_WITH;
 		}
 
 		actionLogger.deleteAllLogs(); // Can't undo past this!
